+++ src/main/java/dinkplugin/notifiers/data/CollectionNotificationData.java
@@ -2,9 +2,20 @@ package dinkplugin.notifiers.data;
 
 import lombok.EqualsAndHashCode;
 import lombok.Value;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 @Value
 @EqualsAndHashCode(callSuper = false)
 public class CollectionNotificationData extends NotificationData {
+
+    @NotNull
     String itemName;
+
+    @Nullable
+    Integer itemId;
+
+    @Nullable
+    Long price;
+
 }